# Conflicts:
#	src/revlanguage/workspace/RbRegister_Dist.cpp

+++ src/core/math/distributions/DistributionGeometric.cpp
@@ -25,14 +25,15 @@ using namespace RevBayesCore;
  * \return Returns a double for the cumulative probability density.
  * \throws Does not throw an error.
  */
-double RbStatistics::Geometric::cdf(int n, double p) {
+double RbStatistics::Geometric::cdf(long n, double p) 
+{
     
     if (p <= 0 || p > 1) 
-        {
+    {
         std::ostringstream s;
         s << "Cannot compute cdf of the Geometric distribution because n = " << n << " is not an integer";
         throw RbException(s.str());
-        }
+    }
     
     if (n < 0.0) 
         return 0.0;
@@ -40,10 +41,10 @@ double RbStatistics::Geometric::cdf(int n, double p) {
         return 1.0;
     
     if (p == 1.0) 
-        { 
+    {
         /* we cannot assume IEEE */
         return n;
-        }
+    }
     n = int( RbMath::log1p(-p) * (n + 1) );
         
     return -RbMath::expm1(n);
@@ -59,7 +60,8 @@ double RbStatistics::Geometric::cdf(int n, double p) {
  * \return Returns a double of the log probability density.
  * \throws Does not throw an error.
  */
-double RbStatistics::Geometric::lnPdf(int n, double p) {
+double RbStatistics::Geometric::lnPdf(long n, double p)
+{
 
     return pdf(n, p, true);
 }
@@ -74,7 +76,8 @@ double RbStatistics::Geometric::lnPdf(int n, double p) {
  * \return Returns a double with the probability density.
  * \throws Does not throw an error.
  */
-double RbStatistics::Geometric::pdf(int n, double p) {
+double RbStatistics::Geometric::pdf(long n, double p) 
+{
 
     return pdf(n, p, false);
 }
@@ -100,16 +103,17 @@ double RbStatistics::Geometric::pdf(int n, double p) {
  * \return Returns the probability density.
  * \throws Does not throw an error.
  */
-double RbStatistics::Geometric::pdf(int n, double p, bool asLog) {
+double RbStatistics::Geometric::pdf(long n, double p, bool asLog)
+{
     
     double prob;
     
     if (p <= 0 || p > 1) 
-        {
+    {
         std::ostringstream s;
         s << "Cannot compute pdf of the Geometric distribution because p = " << p << " is not a probability";
         throw RbException(s.str());
-        }
+    }
     
     if (n < 0 || !RbMath::isFinite(double(n)) || p == 0)
         return ((asLog) ? RbConstants::Double::neginf : 0.0);
@@ -131,20 +135,21 @@ double RbStatistics::Geometric::pdf(int n, double p, bool asLog) {
  * \return Returns the probability density.
  * \throws Does not throw an error.
  */
-int RbStatistics::Geometric::quantile(double q, double p) {
+long RbStatistics::Geometric::quantile(double q, double p)
+{
 
     if (p <= 0 || p > 1) 
-        {
+    {
         std::ostringstream s;
         s << "Cannot compute pdf of the Geometric distribution because q = " << q << " is not an integer";
         throw RbException(s.str());
-        }
+    }
 
     if (p == 1) 
         return 0;
     
     /* add a fuzz to ensure left continuity */
-    return int(ceil(log(q) / RbMath::log1p(- p) - 1 - 1e-7));
+    return long(ceil(log(q) / RbMath::log1p(- p) - 1 - 1e-7));
 }
 
 /*!
@@ -167,8 +172,10 @@ int RbStatistics::Geometric::quantile(double q, double p) {
  *    New York: Springer-Verlag.
  *    Page 480.
  */
-int RbStatistics::Geometric::rv(double p, RevBayesCore::RandomNumberGenerator &rng) {
-    if (!RbMath::isFinite(p) || p <= 0 || p > 1) throw RbException("NaN produced in rgeom");
+long RbStatistics::Geometric::rv(double p, RevBayesCore::RandomNumberGenerator &rng)
+{
+    if (!RbMath::isFinite(p) || p <= 0 || p > 1) 
+        throw RbException("NaN produced in rgeom");
     
     return RbStatistics::Poisson::rv(exp(rng.uniform01()) * ((1 - p) / p),rng);
 }
